+++ lib/phoenix_kit_web/router.ex
@@ -24,6 +24,7 @@ defmodule PhoenixKitWeb.Router do
   pipeline :browser do
     plug :accepts, ["html"]
     plug :fetch_session
+    plug :fetch_live_flash
     plug :put_root_layout, html: PhoenixKit.LayoutConfig.get_root_layout()
     plug :protect_from_forgery
     plug :put_secure_browser_headers